Completed files added
The index, imageSearch and advancedSearch html files contain the html for the main search, image search and advanced search pages.
styles.css contains styling for the html pages.
magnifyingGlass.png is used as the icon in the search bar.
googleLogo.png is taken from the Google branding site and is used on each page.